Phase 5: raw fallback — learn/store/replay non-decodable ("fancy") remotes
READ now falls back to raw capture when the protocol is UNKNOWN, so remotes
that previously only overflowed/refused are learned and replayed.

- ir.cpp capture_raw(): on UNKNOWN, store rawlen-1 timings as 50us "ticks" via
  IrReceiver.compensateAndStoreIRResultInArray() (mark/space compensation,
  leading gap dropped). Over-long frames still refused (WAS_OVERFLOW + a
  defensive rawlen-1 > RAW_MAX_TIMINGS guard), leaving *out untouched.
- ir.cpp ir_send(): raw replay via the uint8_t IrSender.sendRaw(..., 38kHz)
  overload, then IrReceiver.restartAfterSend().
- signal.h: raw[] changed from uint16_t microseconds to uint8_t ticks — the
  receiver's native resolution, and half the RAM/EEPROM footprint.
- storage.cpp: variable-length raw payloads (payload = rawLen tick bytes, RAW
  directory flag, directory length = count; heap-full -> "memory full").
  storage_read reconstructs raw from the directory. Same-length re-store still
  overwrites in place; different-length raw re-store orphans old bytes until
  format() (documented fragmentation trade-off). No EEPROM format-version bump,
  so decoded EEPROMs from Phase 3/4 stay valid.
- Bump to 0.6.0-phase5; retag `raw on|off` help as Phase 6; refresh
  README / project_plan (§3.4 raw format finalized) / session_context.

Verified the IRremote v4.7.1 raw API against library source (uint8_t tick
capture/replay helpers, sendRaw overloads). Off-device: whole sketch
compiles+links clean (-Wall -Wextra) vs mock Arduino/EEPROM/IRremote; raw test
39/39 (capture incl. gap-drop + overflow refusal, EEPROM store/read round-trip,
decoded+raw coexistence, same-/different-length overwrite, memory-full,
sendRaw at 38 kHz, and a serial read->store->dump->send fancy-remote
round-trip); Phase 4 send suite still 33/33.

+++ firmware/signal.h
@@ -25,8 +25,13 @@ struct LearnedSignal {
   uint8_t  flags;
 
   // ---- when type == SIGNAL_RAW ----
+  // Timings are stored as IRremote "ticks" of MICROS_PER_TICK (50 us) each,
+  // one byte per mark/space. That is the receiver's native resolution and
+  // halves the RAM/EEPROM footprint vs. raw microseconds — important on a
+  // 2 KB / 1 KB chip. capture: compensateAndStoreIRResultInArray(); replay:
+  // the uint8_t IrSender.sendRaw() overload (see ir.cpp).
   uint8_t  rawLen;                  // number of timing entries actually used
-  uint16_t raw[RAW_MAX_TIMINGS];    // durations in microseconds
+  uint8_t  raw[RAW_MAX_TIMINGS];    // durations as 50 us ticks (0..255)
 };
 
 // Reset a signal back to empty.
